+++ app/(route)/admin/@modal/(.)notice-create/[shop_id]/_components/schema.ts
