+++ app/(route)/admin/notice-create/[shop_id]/page.tsx
@@ -5,11 +5,9 @@ import CustomTimeInput from "@/app/_components/custom-time-input";
 import CustomTextarea from "@/app/_components/custom-textarea";
 import { yupResolver } from "@hookform/resolvers/yup";
 import { useRouter } from "next/navigation";
-import React, { ElementRef, useEffect, useRef } from "react";
+import React from "react";
 import { useForm } from "react-hook-form";
 import Button from "@/app/_components/button";
-import close from "@/public/icons/close.png";
-import Image from "next/image";
 import { storeNoticeRegisterSchema } from "./schema";
 
 interface SearchParamsProps {
@@ -27,10 +25,6 @@ export default function AddNotice({ searchParams }: SearchParamsProps) {
     formState: { errors },
   } = useForm({ resolver, mode: "onSubmit" });
 
-  function onDismiss() {
-    router.back();
-  }
-
   const onSubmit = async (data: any) => {
     data.startsAt = new Date(data.startsAt).toISOString();
     // const response = await postCreateNotice(shop_id, data);
@@ -42,9 +36,6 @@ export default function AddNotice({ searchParams }: SearchParamsProps) {
       <div className="ml-auto mr-auto mt-[70px] max-w-[964px] pb-20 pl-3 pr-3 pt-[60px] md:pl-8 md:pr-8">
         <div className="mb-6 flex justify-between">
           <h1 className="font-bold text-l">공고 등록</h1>
-          <button onClick={onDismiss}>
-            <Image src={close} alt="close" width={32} height={32} />
-          </button>
         </div>
         <form onSubmit={handleSubmit(onSubmit)}>
           <div className="flex flex-col gap-6">
